refactor: co-locer feature-spesifikke utilities med sine slices
Løser #546 — Flytter feature-koblede utilities ut av src/utils/ og inn
i sine respektive feature-slices.

Server-utilities:
- utils/server/oversikt.ts → features/din-oversikt/server/oversikt.ts

Client-utilities:
- utils/client/dokument.ts  → features/dokumenter/utils/dokument.ts
- utils/client/utbetaling.ts → features/utbetaling/utils/utbetaling.ts
- utils/client/varsler.ts   → features/alert-island/varsler/utils/varsler.ts

src/utils/ inneholder nå kun virkelig delt infrastruktur:
server/: token, fetch, logger, environment, error
client/: amplitude, api, environment, faro
