+++ src/features/alert-island/varsler/Varsler.astro
@@ -1,5 +1,5 @@
 ---
-import { beskjedSingular, buildText, hasVarsler, oppgaveSingular } from "../../../utils/client/varsler";
+import { beskjedSingular, buildText, hasVarsler, oppgaveSingular } from "./utils/varsler";
 import { text } from "../language/varsler";
 import { getAudience, getOboToken } from "../../../utils/server/token";
 import { fetchData } from "../../../utils/server/fetch";
