+++ src/features/din-oversikt/DinOversikt.astro
@@ -4,7 +4,7 @@ import { dinOversiktUrl } from "./utils/urls";
 import MicrofrontendSSR from "../../microfrontends/Microfrontend.astro";
 import Produktkort from "./produktkort/Produktkort";
 import Meldekort from "./meldekort/Meldekort.astro";
-import { getProduktkortList } from "@utils/server/oversikt.ts";
+import { getProduktkortList } from "./server/oversikt.ts";
 import { produktText } from "./produktkort/ProduktText";
 import { getShowDinOversikt } from "./utils/useOversikt.ts";
 import { fetchData } from "@utils/server/fetch";
