+++ src/features/dokumenter/dokument/Dokument.astro
@@ -1,7 +1,7 @@
 ---
 import { dokumentarkivUrl } from "../dokumenterUrls";
 import { BodyShort } from "@navikt/ds-react";
-import { formatDateMonth, setAvsenderMottaker } from "../../../utils/client/dokument";
+import { formatDateMonth, setAvsenderMottaker } from "../utils/dokument";
 import type { DokumentType } from "../dokumenterTypes";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
 import styles from "./Dokument.module.css";
