+++ src/features/utbetaling/enkel-utbetaling/EnkelUtbetaling.astro
@@ -1,7 +1,7 @@
 ---
 import { Heading, BodyShort } from "@navikt/ds-react";
 import { text } from "../language/text";
-import { formatToReadableDate } from "../../../utils/client/utbetaling";
+import { formatToReadableDate } from "../utils/utbetaling";
 import type { UtbetalingData } from "../utbetalingTypes";
 import type { Language } from "../../../language/language";
 import styles from "./EnkelUtbetaling.module.css";
