+++ src/features/utbetaling/utbetaling-list/UtbetalingList.astro
@@ -2,7 +2,7 @@
 import { Language } from "@language/language";
 import { UtbetalingData } from "../utbetalingTypes";
 import { BodyShort, Heading, Tag } from "@navikt/ds-react";
-import { formatToReadableDate } from "../../../utils/client/utbetaling";
+import { formatToReadableDate } from "../utils/utbetaling";
 import { utbetalingsoversiktUrl } from "../utbetalingUrls";
 import styles from "./UtbetalingList.module.css";
 import { ChevronRightIcon as Chevron } from "@navikt/aksel-icons";
